Legendre Series
