+++ implementation/LegendreSeries/test/CMakeLists.txt
@@ -1,0 +1,1 @@
+setupTest(LegendreSeries)
