+++ implementation/LegendreSeries/test/derivative.test.cpp
@@ -2,19 +2,18 @@
 
 #include "catch.hpp"
 
-#include "math/implementation/IteratorLegendreSeries.hpp"
+#include "math/implementation/LegendreSeries.hpp"
 
 extern int testNumber;
 extern std::vector
-<math::implementation::IteratorLegendreSeries
- < std::vector<double>::iterator  > > ils;
+< math::implementation::LegendreSeries
+  < math::implementation::ReferencePolicy, std::vector<double>::iterator > > ils;
 extern std::vector< std::function< double(double) > > diffReferences;
 
 SCENARIO(
-  "The iterator legendre series derivative function"
-  " will return the correct value",
-  "[math], [IteratorLegendreSeries], [derivative]"){
-  GIVEN("an IteratorLegendreSeries and a reference function"){
+  "The legendre series derivative function will return the correct value",
+  "[math], [LegendreSeries], [derivative]"){
+  GIVEN("an LegendreSeries and a reference function"){
     WHEN("queried for a function derivative value"){
       THEN("the values will match"){
         double dx = 2.0/100.0;
@@ -36,17 +35,18 @@ SCENARIO(
 }
 
 SCENARIO(
-  "The iterator legendre series derivative function will evaluate efficiently",
-  "[math], [IteratorLegendreSeries], [benchmark], [hide]"){
-  GIVEN("an IteratorLegendreSeries"){
+  "The legendre series derivative function will evaluate efficiently",
+  "[math], [LegendreSeries], [benchmark], [hide]"){
+  GIVEN("an LegendreSeries"){
     WHEN("queried for a function value"){
       double result;
       THEN("the values will match the reference"){
         std::vector<double> coef;
         coef.resize(65, 0);
         coef.back() = 1;
-        math::implementation::IteratorLegendreSeries
-          < std::vector<double>::iterator  >
+        math::implementation::LegendreSeries
+          < math::implementation::ReferencePolicy,
+            std::vector<double>::iterator  >
           bench(coef.begin(), coef.end());
         result = bench.derivative(0.5);
         REQUIRE( result == Approx(-5.6416));
