+++ implementation/LegendreSeries/test/integrate.test.cpp
@@ -2,36 +2,34 @@
 
 #include "catch.hpp"
 
-#include "math/implementation/IteratorLegendreSeries.hpp"
+#include "math/implementation/LegendreSeries.hpp"
 
 extern int testNumber;
 extern std::function<double(double)> f;
 extern std::function<double(double)> f_int;
 extern std::vector
-< math::implementation::IteratorLegendreSeries< std::vector<double>::iterator > >
-ils;
+< math::implementation::LegendreSeries
+  < math::implementation::ReferencePolicy, std::vector<double>::iterator > > ils;
 extern std::vector< std::function< double(double) > > integralReferences;
 
 SCENARIO(
-  "The iterator legendre series integrate function will return an "
-  " anti-derivative series's",
-  "[math], [IteratorLegendreSeries], [integrate]"){
-  GIVEN("an IteratorLegendreSeries and a reference function anti-derivative"){
+  "The legendre series integrate function will return an anti-derivative series",
+  "[math], [LegendreSeries], [integrate]"){
+  GIVEN("an LegendreSeries and a reference function anti-derivative"){
     WHEN("asked to generate an integral series"){
       THEN("the values will match the reference"){
         {
           LOG(INFO) << "Test " << ++testNumber
                     << ": [integrate] No Errors Expected";
           for (int i = 0 ; i < 11; ++i){
-            auto integralSeries =
-              math::implementation::VectorLegendreSeries(ils[i].integrate(0, 0));
+            auto integralSeries = ils[i].integrate(0, 0);
             double dx = 2.0/100.0;
             double x = -2.0;
             int j = 101;
             double diff = -integralReferences[i](0);
             while (j--){
               REQUIRE( (integralReferences[i](x) + diff) ==
-                       Approx(integralSeries(x)) );
+                       Approx(integralSeries->evaluate(x)) );
               x+= dx;
             }
           }
@@ -40,15 +38,14 @@ SCENARIO(
           LOG(INFO) << "Test " << ++testNumber
                     << ": [integrate] No Errors Expected";
           for (int i = 0 ; i < 11; ++i){
-            auto integralSeries =
-              math::implementation::VectorLegendreSeries(ils[i].integrate(0, 1));
+            auto integralSeries = ils[i].integrate(0, 1);
             double dx = 4.0/100.0;
             double x = -2.0;
             int j = 101;
             double diff = 1 - integralReferences[i](0);
             while (j--){
               REQUIRE( (integralReferences[i](x) + diff) ==
-                       Approx(integralSeries(x)) );
+                       Approx(integralSeries->evaluate(x)) );
               x+= dx;
             }
           }
@@ -57,15 +54,14 @@ SCENARIO(
           LOG(INFO) << "Test " << ++testNumber
                     << ": [integrate] No Errors Expected";
           for (int i = 0 ; i < 11; ++i){
-            auto integralSeries =
-              math::implementation::VectorLegendreSeries(ils[i].integrate(1, 1));
+            auto integralSeries = ils[i].integrate(1, 1);
             double dx = 4.0/100.0;
             double x = -2.0;
             int j = 101;
             double diff = 1 - integralReferences[i](1);
             while (j--){
               REQUIRE( (integralReferences[i](x) + diff) ==
-                       Approx(integralSeries(x)) );
+                       Approx(integralSeries->evaluate(x)) );
               x+= dx;
             }
           }
